@@ -1,7 +1,12 @@
 namespace API.DTOs
 {
     public class VirtualBatteryStateDTO
-    {
-        
+    {        
+        public int RE { get; set; }
+        public int RI { get; set; }
+        public int VE { get; set; }
+        public int VI { get; set; }
+        public int CL { get; set; }
+        public int X { get; set; }
     }
 }
