@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DevComponent } from './dev/dev.component';
+import { HomeComponent } from './home/home.component';
+import { LoginComponent } from './login/login.component';
 
 const routes: Routes = [
-  { path: 'dev', component: DevComponent }
+  {path: "login", component: LoginComponent},
+  {path: "home", component: HomeComponent},
 ];
 
 @NgModule({
